@@ -1,16 +1,10 @@
-import { useRouter } from "next/router";
-import { getTranslation } from "../helper/i18n-helper";
+"use client";
+import {useParams} from "next/navigation";
 
 export default function useTranslate() {
-  const router = useRouter();
-  const { locale } = router;
-
-  /*TODO: inspect this function*/
-  const changeLanguage = async (language: string) => {
-    await router.push("/", "/", { locale: language });
-  };
-
+  const params = useParams();
+  const {locale} = params;
   const isBR = locale === "pt";
 
-  return { t: getTranslation(locale), changeLanguage, isBR };
+  return { isBR };
 }
