@@ -1,5 +1,6 @@
 export { default as Footer } from "./footer";
 export { default as CardContainer } from "./card-container";
-export { default as HeaderImage } from "./header-image";
+// export { default as HeaderImage } from "./header-image";
 export { default as FixedTranslateButton } from "./fixed-translate-btn";
-export { default as FadeInComponent } from "./fadeInComponent";
+// export { default as FadeInComponent } from "./fadeInComponent";
+// export { default as DescriptionCard } from "./description-card";
